Created WrapButton component, tested.
